ex02/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.1.4
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
ecdsa==0.18.0
esptool==4.7.0
executing==2.0.1
idna==3.10
intelhex==2.3.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
load-dotenv==0.1.0
matplotlib-inline==0.1.7
metatrader==0.0.1
MetaTrader5==5.0.4288
nest-asyncio==1.6.0
numpy==1.26.0
packaging==24.0
pandas==2.1.1
parso==0.8.4
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.18.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
reedsolo==1.7.0
requests==2.32.3
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
tzdata==2023.3
urllib3==2.2.3
wcwidth==0.2.13
